CEFOracle

Documentation for CEFOracle.

CEFOracle.H_cef

CEFOracle.H_zeeman

CEFOracle.TAS_resfunc

CEFOracle.blm_dframe

CEFOracle.calc_S_alphabeta

CEFOracle.calc_heatcap

CEFOracle.calc_magnetization

CEFOracle.calc_susceptibility

CEFOracle.cef_eigensystem

CEFOracle.chi2_cef

CEFOracle.dipolar_form_factor

CEFOracle.effective_moment

CEFOracle.partition_function

CEFOracle.population_factor

CEFOracle.ryabov_clm

CEFOracle.single_ion

CEFOracle.spin_operators

CEFOracle.stevens_A

CEFOracle.stevens_EO

CEFOracle.stevens_O

CEFOracle.thermal_average

CEFOracle.transition_matrix_element

CEFOracle.H_cef — Method

Compute the full CEF matrix given a set of Blm parameters

CEFOracle.H_zeeman — Method

Zeeman Hamiltonian given a spin magnitude, an effective and isotropic g-factor and a vector containing the applied magnetic field in Tesla

TODO: define a function method that allows for g to be a tensor, specifically a diagonal tensor in the case the system is aligned in the eigenframe of the g-tensor, i.e. g = [gxx, gyy, gzz]

CEFOracle.TAS_resfunc — Function

Energy transfer dependence of the spectrometer resolution function assumed to be a Voigt profile of fixed Lorentzian width (elastic resolution)

CEFOracle.blm_dframe — Method

blm_dframe(Blm_dict::Dict{String, <:Real})::DataFrame

Given a dictionary of Stevens coefficients of the form Blm -> Value, return a DataFrame with equivalent information.

CEFOracle.calc_S_alphabeta — Method

Simulation of the inelastic neutron scattering x-section. Implementation of eqns (2.42-2.43) of Furrer/Mesot/Strässle and eqn. (8.11) of Boothroyd

CEFOracle.calc_heatcap — Method

Calculation of the Schottky contribution to the specific heat given the energy levels of a CEF model. Implementation of equation 9.25 of Furrer/Messot/Strässle

CEFOracle.calc_magnetization — Method

Calculate the magnetization of a sample given the eigenstates and eigenvalues of the CEF+Zeeman Hamiltonian

Implementation of equation (79) and (82) p.218-219 in Bauer, E., & Rotter, M. (2010). and equivalently equation 9.23 of Furrer/Messot/Strässle

<Jalpha> = sump np <p| J_alpha |p>

CEFOracle.calc_susceptibility — Method

Calculate the single-ion susceptibility of a system given the eigenstates and eigenvalues of the CEF+Zeeman Hamiltonian The result is from first-order perturbation theory

Implementation of the diagonal elements of equation (2.1.18) p.76 in Jensen, J., & Mackintosh, A. R. (1991). Rare earth magnetism. Oxford: Clarendon Press. Equivalently, implementation of equation 9.24 of Furrer/Messot/Strässle

CEFOracle.cef_eigensystem — Function

cef_eigensystem(single_ion::mag_ion, Blm::Dict{String, Real}, Bx::Real, By::Real, Bz::Real:Real, verbose=false)
cef_eigensystem(single_ion::mag_ion, Blm::DataFrame, Bx::Real, By::Real, Bz::Real:Real, verbose=false)

Returns a tuple that contains 1) the full Hamiltonian matrix, $mathcal{H} = mathcal{H}_{rm{CEF}}+mathcal{H}_{rm{Zeeman}}$,

the eigenvalues of the matrix and 3) the eigenvectors of the matrix.

Bx, By and Bz are the values of the applied magnetic field defined in units of Tesla. In addition, the $xyz$ coordinate system is assumed to be parallel to the $abc$ crystal coordinates.

CEFOracle.chi2_cef — Method

given a cef_datasets struct, compute chi^2

CEFOracle.dipolar_form_factor — Method

Magnetic form factor in the dipolar approximation Implementation of equations (6.30) and (6.52) of Boothroyd

CEFOracle.effective_moment — Method

effective_moment(single_ion::mag_ion)::Float64

Calculate the effective magnetic moment in units of Bohr's magneton via $mu_{eff}/mu_{B} = sqrt(J(J+1))$

CEFOracle.partition_function — Method

Calculate the partition function given the eigenvalues of a Hamiltonian Z = sum_p e^(-Ep/kBT)

CEFOracle.population_factor — Method

Calculate Boltzmann thermal occupation factors defined as np = e^(-Ep / kBT)/Z, for all eigenenergies Ep of a Hamiltonian the partition function is Z

CEFOracle.ryabov_clm — Method

Coefficients clm of formula (8) of Ryabov (2009) The Racah operators Otilde are related to the spherical tensors T^lm via Otilde^lpm m = (pm 1)^(m) clm T^l_pm m, This function calculates clm given equations (2), (4) and the fact that clm = alpha / (Nlm * Flm)

CEFOracle.single_ion — Method

single_ion(ion::String)

Given a magnetic ion name, generate a mag_ion object with the following info:

J, lande-factor gJ

Stevens geometric factors theta_l, alpha, beta, gamma

Radial wave functions <r^l> r2, r4, r6

Dipolar magnetic form factor coefficients

Coefficients taken from Rotter Bauer (2010), the McPhase online manual: http://www.mcphase.de/manual5_5/node129.html and Brown's coefficients https://www.ill.eu/sites/ccsl/ffacts/ffactnode1.html to be used in the dipole-approximation equations (6.30) and (6.52) of Boothroyd

CEFOracle.spin_operators — Method

Explicit matrix form of the spin operators for arbitrary spin J

CEFOracle.stevens_A — Method

Factorization of the Stevens Blm parameters defined as Blm = Alm * <r^l> * thetal, where <r^l> is the expectation value of the radial wavefunction of the 4f electron density (tabulated) and thetal are the Stevens factors (as consequence of Wigner-Eckhart) The function below extracts the value of Alm given a B_lm by simple division

CEFOracle.stevens_EO — Method

Calculate the explicit matrix form of the Stevens operator O^l_m Implementation of equation (21) of Ryabov (1999) This definition is consistent with the static list given in the McPhase manual for the Stevens operatos Olm

CEFOracle.stevens_O — Method

Stevens operators as given in http://www.mcphase.de/manual5_5/node131.html

CEFOracle.thermal_average — Method

Calculate the thermal average of an operator O, <O> = sum_p <p|O|p>*np where |p> are the eigenstates of a Hamiltonian and np are the respective populations factors at finite temperature

CEFOracle.transition_matrix_element — Method

Calculate <n|O|m>, where n and m are some eigenvectors of the Hamiltonian and O is a hetmitian operator